(fix) give all simple_app classes their own files
